Character setup forms
This change set creates the character setup component and technically
three forms to set general info, ability scores, and saves.

This is written to the character service and the character object using
an Immutable record. On character update, this data is automatically
saved to the session storage to survive past page reloads.

This sorted out the general mechanic of how to write to the character
object in a controlled way and to ensure events are broadcast when
updates occur.